requirements.txt
fastapi[standard]>=0.104.1
python-jose[cryptography]>=3.3.0
boolia